Separate page component from router and submit functions
